Move SideMenuOptionFactory to quest app module
