+++ android/engine/src/main/res/values/strings.xml
@@ -113,13 +113,6 @@
     <string name="percentage">%</string>
     <string name="logging_out">Please wait.... Logging out</string>
     <string name="session_expired">Session has been expired and must login again</string>
-    <string name="transfer_data">Transfer Data</string>
-    <string name="anc_clients">ANC Clients</string>
-    <string name="all_clients">All Clients</string>
-    <string name="pnc_clients">PNC Clients</string>
-    <string name="family_planning_clients">Family Planning Clients</string>
-    <string name="households">Households</string>
-    <string name="children">Children</string>
     <string name="sex">Sex</string>
     <string name="age">Age</string>
     <string name="dob">DOB</string>
